gravel: ctrl/nodes: split into module
