+++ src/gravel/controllers/bootstrap.py
@@ -1,5 +1,15 @@
 # project aquarium's backend
 # Copyright (C) 2021 SUSE, LLC.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
 import asyncio
 from enum import Enum
@@ -10,8 +20,8 @@ from typing import Optional, List
 from gravel.controllers.config import DeploymentStage
 from gravel.controllers.gstate import gstate
 from gravel.cephadm.cephadm import Cephadm
-from gravel.controllers.nodes import (
-    NodeCantBootstrapError,
+from gravel.controllers.nodes.errors import NodeCantBootstrapError
+from gravel.controllers.nodes.mgr import (
     NodeMgr,
     NodeStageEnum,
     get_node_mgr
